rename: controllerFunctions to common
